#132 get rid of old preview widget
